Crate arrow_arith

Arrow arithmetic and aggregation kernels

Modules

aggregate

Defines aggregations over Arrow arrays.

arity

Defines kernels suitable to perform operations to primitive arrays.

bitwise

boolean

Defines boolean kernels on Arrow BooleanArray’s, e.g. AND, OR and NOT.

numeric

Defines numeric arithmetic kernels on [PrimitiveArray], such as add

temporal

Defines temporal kernels for time and date related functions.