Trait arrow_arith::aggregate::simd::SimdAggregate

pub(super) trait SimdAggregate<T: ArrowNumericType> {
    type ScalarAccumulator;
    type SimdAccumulator;

    // Required methods
    fn init_accumulator_scalar() -> Self::ScalarAccumulator;
    fn init_accumulator_chunk() -> Self::SimdAccumulator;
    fn accumulate_chunk_non_null(
        accumulator: &mut Self::SimdAccumulator,
        chunk: T::Simd
    );
    fn accumulate_chunk_nullable(
        accumulator: &mut Self::SimdAccumulator,
        chunk: T::Simd,
        mask: T::SimdMask
    );
    fn accumulate_scalar(
        accumulator: &mut Self::ScalarAccumulator,
        value: T::Native
    );
    fn reduce(
        simd_accumulator: Self::SimdAccumulator,
        scalar_accumulator: Self::ScalarAccumulator
    ) -> Option<T::Native>;
}

Required Associated Types

type ScalarAccumulator

type SimdAccumulator

Required Methods

fn init_accumulator_scalar() -> Self::ScalarAccumulator

Returns the accumulator for aggregating scalar values

fn init_accumulator_chunk() -> Self::SimdAccumulator

Returns the accumulator for aggregating simd chunks of values

fn accumulate_chunk_non_null( accumulator: &mut Self::SimdAccumulator, chunk: T::Simd )

Updates the accumulator with the values of one chunk

fn accumulate_chunk_nullable( accumulator: &mut Self::SimdAccumulator, chunk: T::Simd, mask: T::SimdMask )

Updates the accumulator with the values of one chunk according to the given vector mask

fn accumulate_scalar( accumulator: &mut Self::ScalarAccumulator, value: T::Native )

Updates the accumulator with one value

fn reduce( simd_accumulator: Self::SimdAccumulator, scalar_accumulator: Self::ScalarAccumulator ) -> Option<T::Native>

Reduces the vector lanes of the simd accumulator and the scalar accumulator to a single value