Module arrow_arith::arithmetic

Defines basic arithmetic kernels for PrimitiveArrays.

These kernels can leverage SIMD if available on your system. Currently no runtime detection is provided, you should enable the specific SIMD intrinsics using RUSTFLAGS="-C target-feature=+avx2" for example. See the documentation here for more information.

Functions

add

Perform left + right operation on two arrays. If either left or right value is null then the result is also null.

add_checked

Perform left + right operation on two arrays. If either left or right value is null then the result is also null.

add_dyn

Perform left + right operation on two arrays. If either left or right value is null then the result is also null.

add_dyn_checked

Perform left + right operation on two arrays. If either left or right value is null then the result is also null.

add_scalar

Add every value in an array by a scalar. If any value in the array is null then the result is also null.

add_scalar_checked

Add every value in an array by a scalar. If any value in the array is null then the result is also null.

add_scalar_checked_dyn

Add every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

add_scalar_dyn

Add every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

divide

Perform left / right operation on two arrays without checking for division by zero or overflow.

divide_and_round 🔒

Divide a decimal native value by given divisor and round the result.

divide_checked

Perform left / right operation on two arrays. If either left or right value is null then the result is also null. If any right hand value is zero then the result of this operation will be Err(ArrowError::DivideByZero).

divide_dyn

Perform left / right operation on two arrays. If either left or right value is null then the result is also null. If any right hand value is zero then the result of this operation will be Err(ArrowError::DivideByZero).

divide_dyn_checked

Perform left / right operation on two arrays. If either left or right value is null then the result is also null. If any right hand value is zero then the result of this operation will be Err(ArrowError::DivideByZero).

divide_dyn_opt

Perform left / right operation on two arrays. If either left or right value is null then the result is also null.

divide_opt

Perform left / right operation on two arrays. If either left or right value is null then the result is also null.

divide_scalar

Divide every value in an array by a scalar. If any value in the array is null then the result is also null. If the scalar is zero then the result of this operation will be Err(ArrowError::DivideByZero).

divide_scalar_checked_dyn

Divide every value in an array by a scalar. If any value in the array is null then the result is also null. If the scalar is zero then the result of this operation will be Err(ArrowError::DivideByZero). The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

divide_scalar_dyn

Divide every value in an array by a scalar. If any value in the array is null then the result is also null. If the scalar is zero then the result of this operation will be Err(ArrowError::DivideByZero). The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

divide_scalar_opt_dyn

Divide every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

get_fixed_point_info 🔒

Returns the precision and scale of the result of a multiplication of two decimal types, and the divisor for fixed point multiplication.

math_checked_divide_op 🔒

Helper function for operations where a valid 0 on the right array should result in an [ArrowError::DivideByZero], namely the division and modulo operations

math_checked_op 🔒

This is similar to math_op as it performs given operation between two input primitive arrays. But the given operation can return Err if overflow is detected. For the case, this function returns an Err.

math_op

Helper function to perform math lambda function on values from two arrays. If either left or right value is null then the output value is also null, so 1 + null is null.

math_safe_divide_op 🔒

modulus

Perform left % right operation on two arrays. If either left or right value is null then the result is also null. If any right hand value is zero then the result of this operation will be Err(ArrowError::DivideByZero).

modulus_dyn

Perform left % right operation on two arrays. If either left or right value is null then the result is also null. If any right hand value is zero then the result of this operation will be Err(ArrowError::DivideByZero).

modulus_scalar

Modulus every value in an array by a scalar. If any value in the array is null then the result is also null. If the scalar is zero then the result of this operation will be Err(ArrowError::DivideByZero).

modulus_scalar_dyn

Modulus every value in an array by a scalar. If any value in the array is null then the result is also null. If the scalar is zero then the result of this operation will be Err(ArrowError::DivideByZero).

multiply

Perform left * right operation on two arrays. If either left or right value is null then the result is also null.

multiply_checked

Perform left * right operation on two arrays. If either left or right value is null then the result is also null.

multiply_dyn

Perform left * right operation on two arrays. If either left or right value is null then the result is also null.

multiply_dyn_checked

Perform left * right operation on two arrays. If either left or right value is null then the result is also null.

multiply_fixed_point

Perform left * right operation on two decimal arrays. If either left or right value is null then the result is also null.

multiply_fixed_point_checked

Perform left * right operation on two decimal arrays. If either left or right value is null then the result is also null.

multiply_fixed_point_dyn

Perform left * right operation on two decimal arrays. If either left or right value is null then the result is also null.

multiply_scalar

Multiply every value in an array by a scalar. If any value in the array is null then the result is also null.

multiply_scalar_checked

Multiply every value in an array by a scalar. If any value in the array is null then the result is also null.

multiply_scalar_checked_dyn

Subtract every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

multiply_scalar_dyn

Multiply every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

negate

Perform - operation on an array. If value is null then the result is also null.

negate_checked

Perform - operation on an array. If value is null then the result is also null.

simd_checked_divide 🔒

Calculates the division operation left / right on two SIMD inputs. The lower-most bits of valid_mask specify which vector lanes are considered as valid.

simd_checked_divide_op 🔒

Creates a new PrimitiveArray by applying simd_op to the left and right input array. If the length of the arrays is not multiple of the number of vector lanes then the remainder of the array will be calculated using scalar_op. Any operation on a NULL value will result in a NULL value in the output.

simd_checked_divide_op_remainder 🔒

Applies op on the remainder elements of two input chunks and writes the result into the remainder elements of result_chunks. The lower-most bits of valid_mask specify which elements are considered as valid.

simd_checked_modulus 🔒

Calculates the modulus operation left % right on two SIMD inputs. The lower-most bits of valid_mask specify which vector lanes are considered as valid.

subtract

Perform left - right operation on two arrays. If either left or right value is null then the result is also null.

subtract_checked

Perform left - right operation on two arrays. If either left or right value is null then the result is also null.

subtract_dyn

Perform left - right operation on two arrays. If either left or right value is null then the result is also null.

subtract_dyn_checked

Perform left - right operation on two arrays. If either left or right value is null then the result is also null.

subtract_scalar

Subtract every value in an array by a scalar. If any value in the array is null then the result is also null.

subtract_scalar_checked

Subtract every value in an array by a scalar. If any value in the array is null then the result is also null.

subtract_scalar_checked_dyn

Subtract every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.

subtract_scalar_dyn

Subtract every value in an array by a scalar. If any value in the array is null then the result is also null. The given array must be a PrimitiveArray of the type same as the scalar, or a DictionaryArray of the value type same as the scalar.