arrow_arith::arity

Function try_binary

pub fn try_binary<A: ArrayAccessor, B: ArrayAccessor, F, O>(
    a: A,
    b: B,
    op: F,
) -> Result<PrimitiveArray<O>, ArrowError>
where
    O: ArrowPrimitiveType,
    F: Fn(A::Item, B::Item) -> Result<O::Native, ArrowError>,

Applies the provided fallible binary operation across a and b.

This will return any error encountered, or collect the results into a [PrimitiveArray]. If any index is null in either a or b, the corresponding index in the result will also be null

Like try_unary the function is only evaluated for non-null indices

Error

Return an error if the arrays have different lengths or the operation is under erroneous