Function arrow_arith::arity::try_binary_mut

pub fn try_binary_mut<T, F>(
    a: PrimitiveArray<T>,
    b: &PrimitiveArray<T>,
    op: F
) -> Result<Result<PrimitiveArray<T>, ArrowError>, PrimitiveArray<T>>
where
    T: ArrowPrimitiveType,
    F: Fn(T::Native, T::Native) -> Result<T::Native, ArrowError>,

Applies the provided fallible binary operation across a and b by mutating the mutable PrimitiveArray a with the results, returning any error. If any index is null in either a or b, the corresponding index in the result will also be null

Like try_unary the function is only evaluated for non-null indices

Mutable primitive array means that the buffer is not shared with other arrays. As a result, this mutates the buffer directly without allocating new buffer.

Error

Return an error if the arrays have different lengths or the operation is under erroneous. This function gives error of original PrimitiveArray a if it is not a mutable primitive array.