arrow_arith::bitwise

Function bitwise_and_scalar

pub fn bitwise_and_scalar<T>(
    array: &PrimitiveArray<T>,
    scalar: T::Native,
) -> Result<PrimitiveArray<T>, ArrowError>
where
    T: ArrowNumericType,
    T::Native: BitAnd<Output = T::Native>,

Perform bitwise and every value in an array with the scalar. If any value in the array is null then the result is also null.