arrow_arith

Module numeric

Defines numeric arithmetic kernels on [PrimitiveArray], such as add

Macros

date ğŸ”’

neg_checked ğŸ”’

neg_wrapping ğŸ”’

op ğŸ”’

Perform an infallible binary operation on potentially scalar inputs

op_ref ğŸ”’

Same as op but with a type hint for the returned array

timestamp ğŸ”’

try_op ğŸ”’

Perform a fallible binary operation on potentially scalar inputs

try_op_ref ğŸ”’

Same as try_op but with a type hint for the returned array

Enums

Op ğŸ”’

An enumeration of arithmetic operations

Traits

DateOp ğŸ”’

Arithmetic trait for date arrays

IntervalOp ğŸ”’

Arithmetic trait for interval arrays

TimestampOp ğŸ”’

Arithmetic trait for timestamp arrays

Functions

add

Perform lhs + rhs, returning an error on overflow

add_wrapping

Perform lhs + rhs, wrapping on overflow for [DataType::is_integer]

arithmetic_op ğŸ”’

Dispatch the given op to the appropriate specialized kernel

date_op ğŸ”’

Perform arithmetic operation on a date array

decimal_op ğŸ”’

Perform arithmetic operation on decimal arrays

div

Perform lhs / rhs

duration_op ğŸ”’

float_op ğŸ”’

Perform an arithmetic operation on floats

integer_op ğŸ”’

Perform an arithmetic operation on integers

interval_op ğŸ”’

Perform arithmetic operation on an interval array

mul

Perform lhs * rhs, returning an error on overflow

mul_wrapping

Perform lhs * rhs, wrapping on overflow for [DataType::is_integer]

neg

Negates each element of array, returning an error on overflow

neg_wrapping

Negates each element of array, wrapping on overflow for [DataType::is_integer]

rem

Perform lhs % rhs

sub

Perform lhs - rhs, returning an error on overflow

sub_wrapping

Perform lhs - rhs, wrapping on overflow for [DataType::is_integer]

timestamp_op ğŸ”’

Perform arithmetic operation on a timestamp array