Module arrow_arith::temporal

Defines temporal kernels for time and date related functions.

Functions

day

Extracts the day of a given temporal primitive array as an array of integers

day_dyn

Extracts the day of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

doy

Extracts the day of year of a given temporal primitive array as an array of integers The day of year that ranges from 1 to 366

doy_dyn

Extracts the day of year of a given temporal array as an array of integers The day of year that ranges from 1 to 366. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

hour

Extracts the hours of a given temporal primitive array as an array of integers within the range of [0, 23].

hour_dyn

Extracts the hours of a given array as an array of integers within the range of [0, 23]. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

microsecond

Extracts the microseconds of a given temporal primitive array as an array of integers

microsecond_dyn

Extracts the microseconds of a given temporal primitive array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

millisecond

Extracts the milliseconds of a given temporal primitive array as an array of integers

millisecond_dyn

Extracts the milliseconds of a given temporal primitive array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

minute

Extracts the minutes of a given temporal primitive array as an array of integers

minute_dyn

Extracts the minutes of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

month

Extracts the month of a given temporal primitive array as an array of integers within the range of [1, 12].

month_dyn

Extracts the month of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

nanosecond

Extracts the nanoseconds of a given temporal primitive array as an array of integers

nanosecond_dyn

Extracts the nanoseconds of a given temporal primitive array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

num_days_from_monday

Extracts the day of week of a given temporal primitive array as an array of integers.

num_days_from_monday_dyn

Extracts the day of week of a given temporal array as an array of integers.

num_days_from_sunday

Extracts the day of week of a given temporal primitive array as an array of integers, starting at Sunday.

num_days_from_sunday_dyn

Extracts the day of week of a given temporal array as an array of integers, starting at Sunday.

quarter

Extracts the quarter of a given temporal primitive array as an array of integers within the range of [1, 4].

quarter_dyn

Extracts the quarter of a given temporal array as an array of integersa within the range of [1, 4]. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

second

Extracts the seconds of a given temporal primitive array as an array of integers

second_dyn

Extracts the seconds of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

using_chrono_tz_and_utc_naive_date_timeDeprecated

Parse the given string into a string representing fixed-offset that is correct as of the given UTC NaiveDateTime. Note that the offset is function of time and can vary depending on whether daylight savings is in effect or not. e.g. Australia/Sydney is +10:00 or +11:00 depending on DST.

week

Extracts the week of a given temporal primitive array as an array of integers

week_dyn

Extracts the week of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

year

Extracts the years of a given temporal primitive array as an array of integers

year_dyn

Extracts the years of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.