arrow_arith
Module aggregate

Defines aggregations over Arrow arrays.

Macros

bit_operation 🔒

Structs

MaxAccumulator 🔒

MinAccumulator 🔒

SumAccumulator 🔒

Constants

PREFERRED_VECTOR_SIZE 🔒

The preferred vector size in bytes for the target platform. Note that the avx512 target feature is still unstable and this also means it is not detected on stable rust.

PREFERRED_VECTOR_SIZE_NON_NULL 🔒

non-nullable aggregation requires fewer temporary registers so we can use more of them for accumulators

Traits

NumericAccumulator 🔒

An accumulator for primitive numeric values.

Functions

aggregate 🔒

Generic aggregation for any primitive type. Returns None if there are no non-null values in array.

aggregate_nonnull_chunk 🔒

aggregate_nonnull_lanes 🔒

aggregate_nonnull_simple 🔒

aggregate_nullable_chunk 🔒

aggregate_nullable_lanes 🔒

bit_and

Returns the bitwise and of all non-null input values.

bit_or

Returns the bitwise or of all non-null input values.

bit_xor

Returns the bitwise xor of all non-null input values.

bool_and

Returns true if all non-null input values are true, otherwise false.

bool_or

Returns true if any non-null input value is true, otherwise false.

max

Returns the maximum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value

max_array

Returns the max of values in the array of ArrowNumericType type, or dictionary array with value of ArrowNumericType type.

max_binary

Returns the maximum value in the binary array, according to the natural order.

max_binary_view

Returns the maximum value in the binary view array, according to the natural order.

max_boolean

Returns the maximum value in the boolean array

max_string

Returns the maximum value in the string array, according to the natural order.

max_string_view

Returns the maximum value in the string view array, according to the natural order.

min

Returns the minimum value in the array, according to the natural order. For floating point arrays any NaN values are considered to be greater than any other non-null value

min_array

Returns the min of values in the array of ArrowNumericType type, or dictionary array with value of ArrowNumericType type.

min_binary

Returns the minimum value in the binary array, according to the natural order.

min_binary_view

Returns the minimum value in the binary view array, according to the natural order.

min_boolean

Returns the minimum value in the boolean array.

min_max_array_helper 🔒

min_max_helper 🔒

Helper to compute min/max of [ArrayAccessor].

min_max_view_helper 🔒

Helper to compute min/max of [GenericByteViewArray<T>]. The specialized min/max leverages the inlined values to compare the byte views. swap_cond is the condition to swap current min/max with the new value. For example, Ordering::Greater for max and Ordering::Less for min.

min_string

Returns the minimum value in the string array, according to the natural order.

min_string_view

Returns the minimum value in the string view array, according to the natural order.

reduce_accumulators 🔒

select 🔒

Helper for branchlessly selecting either a or b based on the boolean m. After verifying the generated assembly this can be a simple if.

sum

Returns the sum of values in the primitive array.

sum_array

Returns the sum of values in the array.

sum_array_checked

Returns the sum of values in the array.

sum_checked

Returns the sum of values in the primitive array.