Module arrow_arith::arity

Kernels for operating on [PrimitiveArray]s

Functions

binary

Allies a binary infallable function to two [PrimitiveArray]s, producing a new [PrimitiveArray]

binary_mut

Applies a binary and infallible function to values in two arrays, replacing the values in the first array in place.

try_binary

Applies the provided fallible binary operation across a and b, returning any error, and collecting the results into a [PrimitiveArray]. If any index is null in either a or b, the corresponding index in the result will also be null

try_binary_mut

Applies the provided fallible binary operation across a and b by mutating the mutable [PrimitiveArray] a with the results, returning any error. If any index is null in either a or b, the corresponding index in the result will also be null

try_binary_no_nulls 🔒

This intentional inline(never) attribute helps LLVM optimize the loop.

try_binary_no_nulls_mut 🔒

This intentional inline(never) attribute helps LLVM optimize the loop.

try_unary

See [PrimitiveArray::try_unary]

try_unary_dict 🔒

A helper function that applies a fallible unary function to a dictionary array with primitive value type.

try_unary_dynDeprecated

Applies a fallible unary function to an array with primitive values.

try_unary_mut

See [PrimitiveArray::try_unary_mut]

unary

See [PrimitiveArray::unary]

unary_dict 🔒

A helper function that applies an infallible unary function to a dictionary array with primitive value type.

unary_dynDeprecated

Applies an infallible unary function to an array with primitive values.

unary_mut

See [PrimitiveArray::unary_mut]