arrow_arith

Module arity

Kernels for operating on [PrimitiveArray]s

Functions

binary

Allies a binary infallable function to two [PrimitiveArray]s, producing a new [PrimitiveArray]

binary_mut

Applies a binary and infallible function to values in two arrays, replacing the values in the first array in place.

create_union_null_buffer 🔒

Computes the union of the nulls in two optional [NullBuffer] which is not shared with the input buffers.

try_binary

Applies the provided fallible binary operation across a and b.

try_binary_mut

Applies the provided fallible binary operation across a and b by mutating the mutable [PrimitiveArray] a with the results.

try_binary_no_nulls 🔒

This intentional inline(never) attribute helps LLVM optimize the loop.

try_binary_no_nulls_mut 🔒

This intentional inline(never) attribute helps LLVM optimize the loop.

try_unary

See [PrimitiveArray::try_unary]

try_unary_mut

See [PrimitiveArray::try_unary_mut]

unary

See [PrimitiveArray::unary]

unary_mut

See [PrimitiveArray::unary_mut]