arrow_arith

Module temporal

Defines temporal kernels for time and date related functions.

Macros

return_compute_error_with 🔒

Enums

DatePart

Valid parts to extract from date/time/timestamp arrays.

Traits

ChronoDateExt 🔒

ExtractDatePartExt 🔒

Implement the specialized functions for extracting date part from temporal arrays.

Functions

date_part

Given an array, return a new array with the extracted DatePart as signed 32-bit integer values.

get_date_time_part_extract_fn 🔒

Returns function to extract relevant DatePart from types like a NaiveDateTime or DateTime.

get_tz 🔒

Extract optional [Tz] from timestamp data types, returning error if called with a non-timestamp type.