arrow_arith

Module temporal

Defines temporal kernels for time and date related functions.

Macros

return_compute_error_with 🔒

Enums

DatePart

Valid parts to extract from date/time/timestamp arrays.

Traits

ChronoDateExt 🔒

ExtractDatePartExt 🔒

Implement the specialized functions for extracting date part from temporal arrays.

Functions

date_part

Given an array, return a new array with the extracted DatePart as signed 32-bit integer values.

date_part_primitive 🔒

Used to integrate new date_part() method with deprecated shims such as hour() and week().

dayDeprecated

Extracts the day of a given temporal primitive array as an array of integers

day_dynDeprecated

Extracts the day of a given temporal array as an array of integers.

doyDeprecated

Extracts the day of year of a given temporal primitive array as an array of integers.

doy_dynDeprecated

Extracts the day of year of a given temporal array as an array of integers.

get_date_time_part_extract_fn 🔒

Returns function to extract relevant DatePart from types like a NaiveDateTime or DateTime.

get_tz 🔒

Extract optional [Tz] from timestamp data types, returning error if called with a non-timestamp type.

hourDeprecated

Extracts the hours of a given temporal primitive array as an array of integers within the range of [0, 23].

hour_dynDeprecated

Extracts the hours of a given array as an array of integers within the range of [0, 23]. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

microsecondDeprecated

Extracts the microseconds of a given temporal primitive array as an array of integers

microsecond_dynDeprecated

Extracts the microseconds of a given temporal primitive array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

millisecondDeprecated

Extracts the milliseconds of a given temporal primitive array as an array of integers

millisecond_dynDeprecated

Extracts the milliseconds of a given temporal primitive array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

minuteDeprecated

Extracts the minutes of a given temporal primitive array as an array of integers

minute_dynDeprecated

Extracts the minutes of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

monthDeprecated

Extracts the month of a given temporal primitive array as an array of integers within the range of [1, 12].

month_dynDeprecated

Extracts the month of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

nanosecondDeprecated

Extracts the nanoseconds of a given temporal primitive array as an array of integers

nanosecond_dynDeprecated

Extracts the nanoseconds of a given temporal primitive array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

num_days_from_mondayDeprecated

Extracts the day of week of a given temporal primitive array as an array of integers.

num_days_from_monday_dynDeprecated

Extracts the day of week of a given temporal array as an array of integers.

num_days_from_sundayDeprecated

Extracts the day of week of a given temporal primitive array as an array of integers, starting at Sunday.

num_days_from_sunday_dynDeprecated

Extracts the day of week of a given temporal array as an array of integers, starting at Sunday.

quarterDeprecated

Extracts the quarter of a given temporal primitive array as an array of integers within the range of [1, 4].

quarter_dynDeprecated

Extracts the quarter of a given temporal array as an array of integersa within the range of [1, 4]. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

secondDeprecated

Extracts the seconds of a given temporal primitive array as an array of integers

second_dynDeprecated

Extracts the seconds of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

weekDeprecated

Extracts the week of a given temporal primitive array as an array of integers

week_dynDeprecated

Extracts the week of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.

yearDeprecated

Extracts the years of a given temporal primitive array as an array of integers

year_dynDeprecated

Extracts the years of a given temporal array as an array of integers. If the given array isn’t temporal primitive or dictionary array, an Err will be returned.